mark data outside climatology ranges not eval
Previous behavior: If a climatology isn't defined for the given time range (and/or depth), then we mark it as qc flag MISSING.
However, the data isn't missing, we just didn't run tests there because they weren't defined for that time/depth.
So mark that point as UNKNOWN/NOT_EVAL instead.