Instead of returning dictionary test results, yield the individual results
This change allows us to better handle additional Streams (that actually
stream) and Stores (that can write based on a generator and not a very
confusing dict structure).

This is a huge change and is not backwards compatible with previous Stream
or Store code. To get the functionality of the old return style, you need
to first collect any results using
`ioos_qc.results.collect_results(results, how='dict')`. See all of the
test changes for examples.

This does NOT affect anyone that was using the QC checks directly or
was using  QcConfig.run() or NcQcConfig.run(), both of which remain
backwards compatible.