Update the spike test differential method
Make sure that np.ma.masked_invalid is called on the diff array in the
differential method to ensure that the mask is a boolean array.
Otherwise, if all values are not masked in the array, the mask
will be a single nomask False value, and this will cause an error in
line 627.

Add to the qartod test_spike_methods test

The test now checks an input array with no missing values for the
differential method, which previously would have failed before this
update.

Resolves: #165